Move soap::tests module into its own file
